requirements.txt
selenium~=4.7.2
HTMLTestRunner-rv==1.0.18
webdriver-manager==3.8.5
